+++ src/main/java/com/tigercel/controller/GroupController.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
@@ -31,43 +32,49 @@ public class GroupController {
     @RequestMapping(value = "", method = GET)
     public String index(Model model) {
 
+        model.addAttribute("menu", "groups");
+        model.addAttribute("title", "分组列表");
+        model.addAttribute("rules", ruleService.findAll());
+
         return "group/index";
     }
 
     @RequestMapping(value = "new", method = POST)
-    public String add(GroupBean gb) {
+    public String add(GroupBean gb, Model model) {
 
         Group group = groupService.findByName(gb.getName());
 
         if(group != null) {
+            model.addAttribute("error", "分组名已存在");
             return "msg/error";
         }
         group = DTOUtil.map(gb, Group.class);
 
-        if(gb.getRuleId() > 0) {
-            groupService.save(group, ruleService.findOne(gb.getRuleId()));
+        if(gb.getRuleName() != null) {
+            groupService.save(group, ruleService.findByName(gb.getRuleName()));
         }
         else {
             groupService.save(group);
         }
-
 
         return "msg/success";
     }
 
     @RequestMapping(value = "edit", method = POST)
     public String edit(GroupBean gb) {
+        Group group = null;
 
-        Group group = groupService.findOne(gb.getId());
+        if(gb.getId() > 0) {
+            group = groupService.findOne(gb.getId());
 
-        DTOUtil.mapTo(gb, group);
-        if(gb.getRuleId() > 0) {
-            groupService.save(group, ruleService.findOne(gb.getRuleId()));
+            DTOUtil.mapTo(gb, group);
+            if(gb.getRuleName() != null) {
+                groupService.save(group, ruleService.findByName(gb.getRuleName()));
+            }
+            else {
+                groupService.save(group);
+            }
         }
-        else {
-            groupService.save(group);
-        }
-
 
         return "msg/success";
     }
@@ -75,16 +82,32 @@ public class GroupController {
     @RequestMapping(value = "delete", method = POST)
     public String delete(GroupBean gb) {
 
-        groupService.delete(gb.getId());
+        if(gb.getId() > 0) {
+            groupService.delete(gb.getId());
+        }
 
         return "msg/success";
     }
 
 
     @RequestMapping(value = "show", method = GET)
-    public String show(PageBean pb, Model model) {
+    public String show(@RequestParam(value = "search", required = false) String search,
+                       PageBean pb, Model model) {
 
-        Page<Group> groups = groupService.findAll(pb);
+        Page<Group> groups = null;
+
+        if(search != null && search.length() != 0) {
+            groups = groupService.findAllByName(pb, search);
+        }
+        else {
+            groups = groupService.findAll(pb);
+        }
+
+        for(Group group : groups.getContent()) {
+            if(group.getRule() != null) {
+                group.setRuleName(group.getRule().getName());
+            }
+        }
 
         model.addAttribute("total", groups.getTotalElements());
         model.addAttribute("rows", groups.getContent());
