+++ src/main/java/com/tigercel/controller/RuleController.java
@@ -12,6 +12,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
@@ -28,17 +29,21 @@ public class RuleController {
     private static Logger logger = LoggerFactory.getLogger(RuleController.class);
 
     @RequestMapping(value = "", method = GET)
-    public String index() {
+    public String index(Model model) {
+
+        model.addAttribute("menu", "rules");
+        model.addAttribute("title", "规则列表");
 
         return "rule/index";
     }
 
     @RequestMapping(value = "new", method = POST)
-    public String add(RuleBean rb) {
+    public String add(RuleBean rb, Model model) {
 
         Rule rule = ruleService.findByName(rb.getName());
 
         if(rule != null) {
+            model.addAttribute("error", "规则名已存在");
             return "msg/error";
         }
         rule = DTOUtil.map(rb, Rule.class);
@@ -49,10 +54,13 @@ public class RuleController {
 
     @RequestMapping(value = "edit", method = POST)
     public String edit(RuleBean rb) {
+        Rule rule = null;
 
-        Rule rule = ruleService.findOne(rb.getId());
-        DTOUtil.mapTo(rb, rule);
-        ruleService.save(rule);
+        if(rb.getId() > 0) {
+            rule = ruleService.findOne(rb.getId());
+            DTOUtil.mapTo(rb, rule);
+            ruleService.save(rule);
+        }
 
         return "msg/success";
     }
@@ -60,19 +68,30 @@ public class RuleController {
     @RequestMapping(value = "delete", method = POST)
     public String delete(RuleBean rb) {
 
-        ruleService.delete(rb.getId());
+        if(rb.getId() > 0) {
+            ruleService.delete(rb.getId());
+        }
 
         return "msg/success";
     }
 
 
     @RequestMapping(value = "show", method = GET)
-    public String show(PageBean pb, Model model) {
-        Page<Rule> rules = ruleService.findAll(pb);
+    public String show(@RequestParam(value = "search", required = false) String search,
+                       PageBean pb, Model model) {
+        Page<Rule> rules = null;
+
+        if(search != null && search.length() != 0) {
+            rules = ruleService.findAllByName(pb, search);
+        }
+        else {
+            rules = ruleService.findAll(pb);
+        }
 
         model.addAttribute("total", rules.getTotalElements());
         model.addAttribute("rows", rules.getContent());
 
         return "jsonTemplate";
     }
+
 }
